INTCMN-108: Refactoring. Under dataservice pkg, use model subpackages for model classes. Version ==> 6.0.0-SNAPSHOT.
