+++ src/main/java/com/blackducksoftware/integration/hub/dataservice/notification/model/ComponentVulnerabilitySummary.java
@@ -21,7 +21,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package com.blackducksoftware.integration.hub.dataservice.notification.item;
+package com.blackducksoftware.integration.hub.dataservice.notification.model;
 
 public class ComponentVulnerabilitySummary {
