+++ src/main/java/com/blackducksoftware/integration/hub/dataservice/notification/model/ComponentVulnerabilitySummaryBuilder.java
@@ -21,7 +21,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package com.blackducksoftware.integration.hub.dataservice.notification.item;
+package com.blackducksoftware.integration.hub.dataservice.notification.model;
 
 import java.util.ArrayList;
 import java.util.List;
