+++ src/main/java/com/blackducksoftware/integration/hub/notification/processor/NotificationProcessor.java
@@ -31,7 +31,7 @@ import java.util.List;
 import java.util.Map;
 import java.util.SortedSet;
 
-import com.blackducksoftware.integration.hub.dataservice.notification.item.NotificationContentItem;
+import com.blackducksoftware.integration.hub.dataservice.notification.model.NotificationContentItem;
 import com.blackducksoftware.integration.hub.exception.HubIntegrationException;
 import com.blackducksoftware.integration.hub.notification.processor.event.NotificationEvent;
