+++ src/main/java/com/blackducksoftware/integration/hub/notification/processor/NotificationSubProcessor.java
@@ -27,7 +27,7 @@ import java.util.Collection;
 import java.util.Map;
 
 import com.blackducksoftware.integration.hub.api.item.MetaService;
-import com.blackducksoftware.integration.hub.dataservice.notification.item.NotificationContentItem;
+import com.blackducksoftware.integration.hub.dataservice.notification.model.NotificationContentItem;
 import com.blackducksoftware.integration.hub.exception.HubIntegrationException;
 import com.blackducksoftware.integration.hub.notification.processor.event.NotificationEvent;
