+++ src/test/java/com/blackducksoftware/integration/hub/notification/processor/ListProcessorCacheTest.java
@@ -35,7 +35,7 @@ import java.util.Map;
 import org.junit.Test;
 
 import com.blackducksoftware.integration.hub.api.policy.PolicyRule;
-import com.blackducksoftware.integration.hub.dataservice.notification.item.PolicyViolationContentItem;
+import com.blackducksoftware.integration.hub.dataservice.notification.model.PolicyViolationContentItem;
 import com.blackducksoftware.integration.hub.notification.processor.event.NotificationEvent;
 
 public class ListProcessorCacheTest {
